hints: separate out hint protobuf into separate package
This commit adds a new package `hintpb` for the protobuf structs used
to store hints in the `system.statement_hints` table. This will avoid
dependency cycles in future commits.

Epic: None

Release note: None
