+++ pkg/sql/hintpb/BUILD.bazel
@@ -1,0 +1,29 @@
+load("@rules_proto//proto:defs.bzl", "proto_library")
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+load("@io_bazel_rules_go//proto:def.bzl", "go_proto_library")
+
+proto_library(
+    name = "hintpb_proto",
+    srcs = ["statement_hint.proto"],
+    strip_import_prefix = "/pkg",
+    visibility = ["//visibility:public"],
+    deps = ["@com_github_gogo_protobuf//gogoproto:gogo_proto"],
+)
+
+go_proto_library(
+    name = "hintpb_go_proto",
+    compilers = ["//pkg/cmd/protoc-gen-gogoroach:protoc-gen-gogoroach_compiler"],
+    importpath = "github.com/cockroachdb/cockroach/pkg/sql/hintpb",
+    proto = ":hintpb_proto",
+    visibility = ["//visibility:public"],
+    deps = ["@com_github_gogo_protobuf//gogoproto"],
+)
+
+go_library(
+    name = "hintpb",
+    srcs = ["statement_hint.go"],
+    embed = [":hintpb_go_proto"],
+    importpath = "github.com/cockroachdb/cockroach/pkg/sql/hintpb",
+    visibility = ["//visibility:public"],
+    deps = ["//pkg/util/protoutil"],
+)
