+++ pkg/sql/hintpb/statement_hint.go
@@ -3,7 +3,7 @@
 // Use of this software is governed by the CockroachDB Software License
 // included in the /LICENSE file.
 
-package hints
+package hintpb
 
 import "github.com/cockroachdb/cockroach/pkg/util/protoutil"
