+++ pkg/sql/hints/BUILD.bazel
@@ -1,14 +1,8 @@
-load("@rules_proto//proto:defs.bzl", "proto_library")
-load("@io_bazel_rules_go//proto:def.bzl", "go_proto_library")
 load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "hints",
-    srcs = [
-        "hint_cache.go",
-        "statement_hint.go",
-    ],
-    embed = [":hints_go_proto"],
+    srcs = ["hint_cache.go"],
     importpath = "github.com/cockroachdb/cockroach/pkg/sql/hints",
     visibility = ["//visibility:public"],
     deps = [
@@ -25,6 +19,7 @@ go_library(
         "//pkg/sql/catalog/descpb",
         "//pkg/sql/catalog/descs",
         "//pkg/sql/catalog/systemschema",
+        "//pkg/sql/hintpb",
         "//pkg/sql/rowenc",
         "//pkg/sql/sem/tree",
         "//pkg/sql/sessiondata",
@@ -34,30 +29,12 @@ go_library(
         "//pkg/util/hlc",
         "//pkg/util/log",
         "//pkg/util/metamorphic",
-        "//pkg/util/protoutil",
         "//pkg/util/retry",
         "//pkg/util/startup",
         "//pkg/util/stop",
         "//pkg/util/syncutil",
         "@com_github_cockroachdb_errors//:errors",
     ],
-)
-
-go_proto_library(
-    name = "hints_go_proto",
-    compilers = ["//pkg/cmd/protoc-gen-gogoroach:protoc-gen-gogoroach_compiler"],
-    importpath = "github.com/cockroachdb/cockroach/pkg/sql/hints",
-    proto = ":hints_proto",
-    visibility = ["//visibility:public"],
-    deps = ["@com_github_gogo_protobuf//gogoproto"],
-)
-
-proto_library(
-    name = "hints_proto",
-    srcs = ["statement_hint.proto"],
-    strip_import_prefix = "/pkg",
-    visibility = ["//visibility:public"],
-    deps = ["@com_github_gogo_protobuf//gogoproto:gogo_proto"],
 )
 
 go_test(
@@ -80,6 +57,7 @@ go_test(
         "//pkg/server",
         "//pkg/sql/catalog",
         "//pkg/sql/catalog/descs",
+        "//pkg/sql/hintpb",
         "//pkg/sql/randgen",
         "//pkg/sql/stats",
         "//pkg/testutils",
