+++ pkg/sql/hints/hint_cache.go
@@ -25,6 +25,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/descpb"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/descs"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/systemschema"
+	"github.com/cockroachdb/cockroach/pkg/sql/hintpb"
 	"github.com/cockroachdb/cockroach/pkg/sql/rowenc"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/sessiondata"
@@ -392,7 +393,7 @@ func (c *StatementHintsCache) GetGeneration() int64 {
 // retrieving them.
 func (c *StatementHintsCache) MaybeGetStatementHints(
 	ctx context.Context, statementFingerprint string,
-) (hints []StatementHint, ids []int64) {
+) (hints []hintpb.StatementHint, ids []int64) {
 	hash := fnv.New64()
 	_, err := hash.Write([]byte(statementFingerprint))
 	if err != nil {
@@ -450,7 +451,7 @@ func (c *StatementHintsCache) maybeWaitForRefreshLocked(
 // released while reading from the db, and then reacquired.
 func (c *StatementHintsCache) addCacheEntryLocked(
 	ctx context.Context, statementHash int64, statementFingerprint string,
-) (hints []StatementHint, ids []int64) {
+) (hints []hintpb.StatementHint, ids []int64) {
 	c.mu.AssertHeld()
 
 	// Add a cache entry that other queries can find and wait on until we have the
@@ -515,7 +516,7 @@ func (c *StatementHintsCache) getStatementHintsFromDB(
 		datums := it.Cur()
 		rowID := int64(tree.MustBeDInt(datums[0]))
 		fingerprint := string(tree.MustBeDString(datums[1]))
-		hint, err := NewStatementHint([]byte(tree.MustBeDBytes(datums[2])))
+		hint, err := hintpb.NewStatementHint([]byte(tree.MustBeDBytes(datums[2])))
 		if err != nil {
 			return err
 		}
@@ -542,7 +543,7 @@ type cacheEntry struct {
 	// be duplicate entries in the fingerprints slice.
 	// TODO(drewk): consider de-duplicating the fingerprint strings to reduce
 	// memory usage.
-	hints        []StatementHint
+	hints        []hintpb.StatementHint
 	fingerprints []string
 	ids          []int64
 }
@@ -551,7 +552,7 @@ type cacheEntry struct {
 // fingerprint, or nil if they don't exist. The results are in order of row ID.
 func (entry *cacheEntry) getMatchingHints(
 	statementFingerprint string,
-) (hints []StatementHint, ids []int64) {
+) (hints []hintpb.StatementHint, ids []int64) {
 	for i := range entry.hints {
 		if entry.fingerprints[i] == statementFingerprint {
 			hints = append(hints, entry.hints[i])
